fix some gcc warnings on Windows
the return of APR_EGENERAL is to bypass a theoretical reference to an 
unset variable (sizelo) in a should-not-occur path, following an
existing example


git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@1083227 13f79535-47bb-0310-9956-ffa450edef68

+++ user/win32/userinfo.c
@@ -46,7 +46,7 @@ static void get_sid_string(char *buf, apr_size_t blen, apr_uid_t id)
     sa  =  (DWORD)(psia->Value[1])        + ((DWORD)(psia->Value[0]) <<  8);
     if (sa) {
         slen = apr_snprintf(buf, blen, "S-%d-0x%04x%08x",
-                            SID_REVISION, sa, nsa);
+                            SID_REVISION, (unsigned int)sa, (unsigned int)nsa);
     } else {
         slen = apr_snprintf(buf, blen, "S-%d-%lu",
                             SID_REVISION, nsa);
